Add files via upload
The Solvee application as a means of submitting aspirations and online complaints is a system built to assist the community in making complaints and helping village and sub-district governments to manage complaints submitted by the community. This system is built on a website platform used by 4 users, namely admin, community, village government and sub-district government.